09.20.24 Fully Functioning Version of request_form.py
The code is confirmed to be capable of setting up a secret word for specified administrators, with each person having their own unique word combination.

These words all update for each administrator every three months, and the administrators receive an email from the system on what their new secret word is. 

The email sent by perrequestform@gmail.com gets deleted on the Gmail end for data security, and only the administrators are still able to have access to their secret word. 

The secret word for the administrators is available in Salesforce on the backend, hidden away from the page layout.